Switch to using the native pg_catalog schema to generate definitions
This is much less reliant on heavy dependencies like a DBIx::Class
schema and the various Perl deps required to successfully load it at
runtime.  I always knew this was a better solution that I hoped to use
one day.  I have that excuse now as I prepared to add another app to the
mix of our reporting database.